SLING-7699 Adding code to embed into Resource object
removed class based, added junit
